Support to reboot into UART download mode
without any external wiring.

This patch introduces the new method
Esp.rebootIntoUartDownloadMode()

When the user calls this method the ESP8266 reboots into the UART
download mode. In this mode the user can use esptool.py to flash a new
firmware file.
This patch tries to support even the esptool.py flasher stub mode.
Therefore option --no-stub would no longer be required.

The implementation is based on the original implementation in the
boot ROM. Some parts of the original implementation can be found in
https://github.com/twischer/xtensa-subjects/blob/master/reversed/bootrom.c

Signed-off-by: Timo Wischer <twischer@freenet.de>